+++ packages/react-together/src/components/Cursors/Cursors.tsx
@@ -1,30 +1,46 @@
 import { useCursors } from '../../hooks'
 import { UseCursorsOptions } from '../../hooks/useCursors'
 import './Cursors.css'
-import UserCursor, { UserCursorOptions, UserCursorProps } from './UserCursor'
+import _UserCursor, { UserCursorOptions, UserCursorProps } from './UserCursor'
 
 interface CursorsComponents {
   UserCursor?: (props: UserCursorProps) => React.ReactNode
 }
 
-interface CursorsProps extends UserCursorOptions {
-  useCursorsOptions?: UseCursorsOptions
+interface CursorsProps extends UserCursorOptions, UseCursorsOptions {
   components?: CursorsComponents
 }
 
-export default function Cursors(options: CursorsProps = {}) {
-  const { useCursorsOptions, components } = options
+export default function Cursors({
+  // useCursorsOptions
+  omitMyValue,
+  deleteOnLeave,
+  throttleDelay,
+  // UserCursorOptions
+  transitionDuration,
+  getUserColor,
+  // CursorsComponents
+  components
+}: CursorsProps = {}) {
+  const { allCursors } = useCursors({
+    omitMyValue,
+    deleteOnLeave,
+    throttleDelay
+  })
 
-  const { allCursors } = useCursors(useCursorsOptions)
-
-  const UserCursorComponent = components?.UserCursor ?? UserCursor
+  const UserCursor = components?.UserCursor ?? _UserCursor
 
   return (
     <div className="cursors-container">
       {Object.entries(allCursors).map(
         ([userId, cursor]) =>
           cursor && (
-            <UserCursorComponent key={userId} userId={userId} {...cursor} />
+            <UserCursor
+              key={userId}
+              userId={userId}
+              {...cursor}
+              {...{ transitionDuration, getUserColor }}
+            />
           )
       )}
     </div>