+++ website/src/components/demo/index.ts
@@ -1,25 +1,9 @@
-import { ChatDemo } from './ChatDemo'
-import { ConnectedUsersDemo } from './ConnectedUsersDemo'
 import CountButtonTogether from './CountButtonTogether'
 import HeroDemo from './HeroDemo'
-import { HoverHighlighterDemo } from './HoverHighlighterDemo'
-import { MeditationBell } from './MeditationBell'
-import { SessionManagerDemo } from './SessionManagerDemo'
-import { UseChatDemo } from './UseChatDemo'
-import { UseConnectedUsersDemo } from './UseConnectedUsersDemo'
-import { UseStateTogetherWPUVDemo } from './UseStateTogetherWPUVDemo'
-export * from './antdesign'
-export * from './primereact'
 
-export {
-  ChatDemo,
-  ConnectedUsersDemo,
-  CountButtonTogether,
-  HeroDemo,
-  HoverHighlighterDemo,
-  MeditationBell,
-  SessionManagerDemo,
-  UseChatDemo,
-  UseConnectedUsersDemo,
-  UseStateTogetherWPUVDemo,
-}
+export * from './antdesign'
+export * from './hooks'
+export * from './primereact'
+export * from './reactTogether'
+
+export { CountButtonTogether, HeroDemo }