+++ website/src/pages/Documentation/components/reactTogether/index.ts
@@ -1,0 +1,13 @@
+import ChatDocumentationPage from './Chat'
+import ConnectedUsersDocumentationPage from './ConnectedUsers'
+import HoverHighlighterDocumentationPage from './HoverHighlighter'
+import ReactTogetherDocumentationPage from './ReactTogether'
+import SessionManagerDocumentationPage from './SessionManager'
+
+export {
+  ChatDocumentationPage,
+  ConnectedUsersDocumentationPage,
+  HoverHighlighterDocumentationPage,
+  ReactTogetherDocumentationPage,
+  SessionManagerDocumentationPage,
+}