+++ website/src/pages/Documentation/hooks/index.ts
@@ -1,25 +1,31 @@
+import UseAllNicknamesDocumentationPage from './useAllNicknames'
 import UseChatDocumentationPage from './useChat'
 import UseConnectedUsersDocumentationPage from './useConnectedUsers'
 import UseCreateRandomSessionDocumentationPage from './useCreateRandomSession'
+import UseCursorsDocumentationPage from './useCursors'
 import UseFunctionTogetherDocumentationPage from './useFunctionTogether'
 import UseHoveringUsersDocumentationPage from './useHoveringUsers'
 import UseIsTogetherDocumentationPage from './useIsTogether'
 import UseJoinUrlDocumentationPage from './useJoinUrl'
 import UseLeaveSessionDocumentationPage from './useLeaveSession'
 import UseMyIdDocumentationPage from './useMyId'
+import UseNicknameDocumentationPage from './useNicknames'
 import UseStateTogetherDocumentationPage from './useStateTogether'
 import UseStateTogetherWithPerUserValuesDocumentationPage from './useStateTogetherWithPerUserValues'
 
 export {
+  UseAllNicknamesDocumentationPage,
   UseChatDocumentationPage,
   UseConnectedUsersDocumentationPage,
   UseCreateRandomSessionDocumentationPage,
+  UseCursorsDocumentationPage,
   UseFunctionTogetherDocumentationPage,
   UseHoveringUsersDocumentationPage,
   UseIsTogetherDocumentationPage,
   UseJoinUrlDocumentationPage,
   UseLeaveSessionDocumentationPage,
   UseMyIdDocumentationPage,
+  UseNicknameDocumentationPage,
   UseStateTogetherDocumentationPage,
   UseStateTogetherWithPerUserValuesDocumentationPage,
 }